removendo arquivos sem uso e tirando abreviacao
