+++ GerenciamentoClientes/GerenciamentoClientes.csproj
@@ -11,13 +11,11 @@
   <ItemGroup>
     <Compile Remove="Form2.cs" />
     <Compile Remove="Form2.Designer.cs" />
-    <Compile Remove="Tela_Cadastro.cs" />
     <Compile Remove="Tela_Inicial.cs" />
   </ItemGroup>
 
   <ItemGroup>
     <EmbeddedResource Remove="Form2.resx" />
-    <EmbeddedResource Remove="Tela_Cadastro.resx" />
     <EmbeddedResource Remove="Tela_Inicial.resx" />
   </ItemGroup>
